set context and identity
Mumble spatializes based on users with shared contexts. Identity strings
provide information on users that can be useful on server scripts.

With this patch, players in the same channel but different arenas should
no longer hear positional audio for each other.

For now, the identity string is just the player name, but the team color could
be added in the future to support automatically placing users in the
appropriate channels.

Signed-off-by: Quentin Young <qlyoung@qlyoung.net>